GraphMatFun

GraphMatFun.Compgraph

GraphMatFun.Compgraph

GraphMatFun.Degopt

GraphMatFun.LangC_MKL

GraphMatFun.LangC_OpenBLAS

GraphMatFun.LangJulia

GraphMatFun.LangMatlab

GraphMatFun.add_ldiv!

GraphMatFun.add_lincomb!

GraphMatFun.add_mult!

GraphMatFun.add_output!

GraphMatFun.add_sum!

GraphMatFun.adjust_for_errtype!

GraphMatFun.clear_outputs!

GraphMatFun.compress_graph!

GraphMatFun.compress_graph_dangling!

GraphMatFun.compress_graph_output_cleaning!

GraphMatFun.compress_graph_passthrough!

GraphMatFun.compress_graph_redundant!

GraphMatFun.compress_graph_trivial!

GraphMatFun.compress_graph_zero_coeff!

GraphMatFun.compute_bwd_theta_exponential

GraphMatFun.compute_fwd_theta

GraphMatFun.del_node!

GraphMatFun.del_output!

GraphMatFun.eval_graph

GraphMatFun.eval_jac

GraphMatFun.eval_runerr

GraphMatFun.export_compgraph

GraphMatFun.extract_sums

GraphMatFun.gen_code

GraphMatFun.get_all_cref

GraphMatFun.get_children

GraphMatFun.get_coeffs

GraphMatFun.get_degopt_crefs

GraphMatFun.get_polynomial

GraphMatFun.get_polynomial_coefficients

GraphMatFun.get_sorted_keys

GraphMatFun.get_topo_order

GraphMatFun.get_topo_order_degopt

GraphMatFun.graph_bbc_exp

GraphMatFun.graph_bbcs_cheb_exp

GraphMatFun.graph_degopt

GraphMatFun.graph_denman_beavers

GraphMatFun.graph_exp_native_jl

GraphMatFun.graph_exp_native_jl_degopt

GraphMatFun.graph_horner

GraphMatFun.graph_horner_degopt

GraphMatFun.graph_monomial

GraphMatFun.graph_monomial_degopt

GraphMatFun.graph_newton_schulz

GraphMatFun.graph_newton_schulz_degopt

GraphMatFun.graph_ps

GraphMatFun.graph_ps_degopt

GraphMatFun.graph_rational

GraphMatFun.graph_sastre_exp

GraphMatFun.graph_sastre_poly

GraphMatFun.graph_sastre_yks_degopt

GraphMatFun.graph_sid_exp

GraphMatFun.grow!

GraphMatFun.has_identity_lincomb

GraphMatFun.has_trivial_nodes

GraphMatFun.import_compgraph

GraphMatFun.merge_graphs

GraphMatFun.opt_gauss_newton!

GraphMatFun.opt_linear_fit!

GraphMatFun.rename_node!

GraphMatFun.scale!

GraphMatFun.set_coeffs!

GraphMatFun.solve_linlsqr!

GraphMatFun.square!

LinearAlgebra.normalize!

GraphMatFun.Compgraph — Type

graph=Compgraph(T::Type=ComplexF64)

Creates an empty computation graph of with coefficients of type T.

GraphMatFun.Compgraph — Method

graph=Compgraph(T,orggraph::Compgraph)

Converts a graph such that the coefficients become type T. Note that T can be a Number but can be other objects with defined operations, or just Any.

GraphMatFun.Degopt — Method

degopt=Degopt(x,y)
degopt=Degopt(HA,HB,y)
degopt=Degopt(graph)

struct Degopt{T}
    x::Vector{Tuple{Vector{T},Vector{T}}}
    y::Vector{T}
end

Creates an object representing a degree optimal polynomial, i.e., the coefficients in an evaluation scheme which maximizes the degree for a fixed number of multiplications. The object represents the coefficients in

A0=I
A1=A
A2=(x *A0+x *A1)(x *A0+x *A1)
A4=(x *A0+x *A1+x*A2)(x *A0+x *A1+x *A2)
A8=(x *A0+x *A1+x*A2)(x *A0+x *A1+x *A2)
 ..

and

Out=y*A0+y*A1+y*A4+y*A8+y*B4...

The x-values are given in the argument x, which is a Vector{Tuple{Vector{T},Vector{T}}}, containing the elements of each sum. The y-vector contains the elements to form the output. In the constructor with HA, HB and y the elements of x are stored as matrices.

The coefficients in graphs generated by graph_degopt can be recovered by using Degopt(graph).

GraphMatFun.LangC_MKL — Type

 LangC_MKL(gen_main::Bool)

Code generation in C using MKL implementation of BLAS.

GraphMatFun.LangC_OpenBLAS — Type

 LangC_OpenBLAS(gen_main::Bool)

Code generation in C using OpenBLAS implementation of BLAS.

GraphMatFun.LangJulia — Type

LangJulia(overwrite_input=true,inline=true,dot_fusing=true)

Code generation in julia language, with optional overwriting of input, inlining the function and optional usage of dot fusion.

GraphMatFun.LangMatlab — Type

LangMatlab()

Code generation for the Matlab language.

GraphMatFun.add_ldiv! — Method

add_ldiv!(graph,node,p1,p2)

The operation inv(p1)*p2 is added to the graph. The result is stored in node node.

GraphMatFun.add_lincomb! — Method

add_lincomb!(graph,node,α1,p1,α2,p2)

The operation α1*p1+α2*p2 is added to the graph. The result is stored in node node.

GraphMatFun.add_mult! — Method

add_mult!(graph,node,p1,p2)

Adds a multiplication of node p1 and p2 to the graph. The result is stored in node node.

GraphMatFun.add_output! — Method

add_output!(graph,node)

Adds an output node to the graph.

GraphMatFun.add_sum! — Function

crefs=add_sum!(graph,node,c,nodelist,base_name=node)

Adds a linear combination of more than two nodes, given in nodelist::Vector, with coefficients given in c. The base_name is temporary variables for the summing. The sum is stored in node.

Returns cref list with references.

GraphMatFun.adjust_for_errtype! — Method

adjust_for_errtype!(A, b, objfun_vals, errtype)

Adjusts the matrix A and vector b to errtype. A is a matrix, e.g., the Jacobian or system matrix in a least squares problem. b is a vector, e.g., the residuals or the right-hand side in a lieast squares problem. objfun_vals is a vector of objective function values, and errtype can be :abserr or :relerr, i.e., absolute error or relative error.

GraphMatFun.clear_outputs! — Method

clear_outputs!(graph)

Clears the list of output nodes of the graph. This function does not remove the node nodes from the graph.

GraphMatFun.compress_graph! — Function

compress_graph!(graph,cref=[];verbose=false)

Searches for nodes which can be removed or reorganized without changing the function it represents. Corresponding references in the cref-vector are removed.

GraphMatFun.compress_graph_dangling! — Function

compress_graph_dangling!(graph,cref=[];verbose=false)

Removes dangling nodes in the graph. If a node is not used anywhere and it is not in the output, it can safely be removed.

GraphMatFun.compress_graph_output_cleaning! — Function

compress_graph_output_cleaning!(graph,cref=[];verbose=false)

Checks if the output is computed from the linear combination that can be compressed.

GraphMatFun.compress_graph_passthrough! — Function

compress_graph_passthrough!(graph,cref=[];verbose=false);

Identifies lincombs that have coefficients (0 1) or (1 0) which correspond to identity operations. It redirect appropriately.

GraphMatFun.compress_graph_redundant! — Function

compress_graph_redundant!(graph,compress_lincomb=false,cref=[];verbose=false)

Removes from the graph redundant nodes, that is, nodes that repeat a computaion that is already present in the graph. Nodes corresponding to a linear combination are removed only if the coefficients are the same and compress_lincomb is set to true.

GraphMatFun.compress_graph_trivial! — Function

compress_graph_trivial!(graph,cref=[];verbose=false)

Removes from the graph trivial nodes, that is, multiplications by the identity or linear systems whose coefficient is the identity matrix.

GraphMatFun.compress_graph_zero_coeff! — Function

compress_graph_zero_coeff!(graph,cref=[];droptol=0;verbose=false)

Searches for linear combinations with zero coeff value and tries to compress by redirecting node references.

GraphMatFun.compute_bwd_theta_exponential — Method

(e_bwd,theta)=compute_bwd_theta_exponential(graph::Compgraph{T},
                                            nterms=100,
                                            ndigits=100,
                                            coefftype=T,
                                            tolerance=eps(coefftype)/2,
                                            theta_init=big"0.1") where T

Bound on relative backward error of the exponential with corresponding theta.

The first output $e_{bwd}$ is the function that computes a bound on the relative backward error of the polynomial underlying the computational graph graph, seen as an approximant to the exponential. In other words, for the underlying polynomial $p$ function returns a bound on |δ| where δ is such that exp(z+δ) = p(z).

The bound is computed by means of the identity δ = log((exp(-z) p(z)-1)+1). The code computes the series expansion of the right-hand side of the equation, truncates it to the first nterms coefficients, bounds each coefficient of the ensuing polynomial with its absolute value. The computation uses ndigits digits of precision.

The second output $θ$ is the largest positive real number such that $e_{bwd}(θ) ≤ tolerance$, where $tolerance$ is by default the unit roundoff of the data type coefftype, which in turn defaults to the type of the coefficients of graph.

The value of $theta$ is approximated by using the built-in function fzero with starting value set to theta_init. By default this root-finding procedure uses high-precision arithmetic.

GraphMatFun.compute_fwd_theta — Method

(e_fwd,theta)=compute_fwd_theta(graph::Compgraph{T},f;
                                coefftype=T,
                                tolerance=eps(coefftype)/2
                                theta_init=big"0.1") where T

Return relative forward error and corresponding theta for approximation to f.

The first output $e_{fwd}$ is the function that computes the relative forward error$e_{fwd}(z)=|f(z) - p(z)| / |z|$, where $p$ is the polynomial underlying the computational graph graph. This is meaningful only if $p$ is approximates $f$ in some sense.

The second output $θ$ is the largest positive real number such that $e_{fwd}(θ) ≤ tolerance$, where $tolerance$ is by default the unit roundoff of the data type coefftype, which in turn defaults to the type of the coefficients of graph.

The value of $θ$ is approximated by using the built-in function fzero with starting value set to theta_init. By default, this root-finding procedure uses high-precision arithmetic.

GraphMatFun.del_node! — Method

del_node!(graph,node)

Deletes a node from the graph, and all data associated with it, except for graph.outputs.

GraphMatFun.del_output! — Method

del_output!(graph,node)

Removes the output node from the list of output nodes of the graph. This function does not remove the node from the graph.

GraphMatFun.eval_graph — Method

result=eval_graph(graph, x; vals=nothing,
                  input=:A, output=size(graph.outputs,1),
                  comporder=nothing)

Evaluates a graph in the value x which is typically a scalar value or a matrix. If x is a Vector, the values will be evaluated elementwise.

The comporder is a Vector of nodes specifying in which order the graph should be computed. By default get_topo_order is used.

The output is an Int specifying which node should be considered as output. The output node is graph.outputs[output].

The vals is used to inspect contents other than the output inside the graph. Typically vals is a Dict. It will be modified to contain the computed nodes of the graph. If we wish to inspect node X3, we can initiate an empty dict as input:

julia> vals=Dict{Symbol,Any}();
julia> eval_graph(graph,A,vals=vals);
julia> vals[:X3]

GraphMatFun.eval_jac — Method

J=eval_jac(graph,x, cref; vals=nothing,
           input=:A, output=size(graph.outputs,1))

Computes Jacobian J = dZ(x_i)/dc, with respect to the coefficients given in the vector cref, and points x[1],...,x[end]. See eval_graph for description of vals and output.

GraphMatFun.eval_runerr — Method

err=eval_runerr(graph, x; vals=nothing, relerrs=nothing,
                       input=:A,output=size(graph.outputs,1),
                       mode=:bounds,
                       add_relerr=eps())

Provides the running error of the graph evaluated in x. See eval_graph for meaning of vals and output. The kwarg relerrs is an anologous variable for the running error estimates in each node. The kwarg mode can be :bounds, :rand, :estimate, specifying if the code should make estimates or bounds, or random error within the bound.

GraphMatFun.export_compgraph — Method

export_compgraph(graph, fname; main_output=nothing,
                 order=get_topo_order(graph)[1],
                 fun="", dom="", err="", genby="",
                 descr="",
                 user=current user)

Exports the graph to the file fname using the computation graph format (cgr). These kwargs are strings or values that are stored as comments in the file

descr of the graph

fun function it approximates

dom domain it approximates

err error in this domain

genby script that generated this file

user a username that created the file

These kwarg influence how the graph is stored:

order specifies an order the graph nodes are stored

main_output is a Symbol specifying what is the output.

CGR format:

Every line corresponds to a node / operation. The syntax is matlab compatible, although gen_code produces faster matlab code.

GraphMatFun.extract_sums — Method

v=extract_sums(graph)

::Vector{Tuple{Vector{Float64},Vector{Symbol},Vector{Symbol}}}`

Returns a representation of sums in graph which may potentially be merged. The vector sums contains a tuple for each of these sums. The three entries of the tuple are:

a vector of Float64 values that represent the coefficients of the summands;

a vector of Symbols that correspond to the summands; and

a vector of intermediate Symbols (i.e, nodes) that can be merged.

The first two vectors have the same number of entries, one for each element that can be merged in the sum.

GraphMatFun.gen_code — Method

gen_code(fname,graph; priohelp=Dict{Symbol,Float64}(),
         lang=LangJulia(),funname="dummy",precomputed_nodes=[:A])

Generates the code for the graph in the language specified in lang and writes it into the file fname. The string funname is the function name. Topological order of the nodes is comptued using get_topo_order and priohelp can be used to influence the order. The nodes listed in precomputed_nodes are viewed as inputs, and code to compute these nodes are not computed.

Currently supported languages: LangC_MKL, LangC_OpenBLAS, LangJulia, LangMatlab, LangDegoptJulia.

GraphMatFun.get_all_cref — Method

v=get_all_cref(graph)

Returns a list with references to all coefficients in the graph. The list is sorted.

GraphMatFun.get_children — Method

 children=get_children(graph,node)

Returns the (direct) children of node.

GraphMatFun.get_coeffs — Function

x=get_coeffs(graph, cref=get_all_cref(graph))

Gets the coefficient values for the coefficients specified in cref::Vector.

GraphMatFun.get_degopt_crefs — Method

(x,z)=get_degopt_crefs(k)
(x,z)=get_degopt_crefs(graph)

Retruns crefs related to graph_degopt. Specifically x is a Vector{Tuple{Vector{Tuple{Symbol,Int}},Vector{Tuple{Symbol,Int}}}} such that x[2][1] corresponds to the coefficients of the left hand side of the multiplication

B2=(α_2_1 *I + α_2_2 *A)(β_2_1 *I + β_2_2 *A)

i.e., the crefs corresponding to [α_2_1, α_2_2]. See graph_degopt. Hence, get_coeffs(graph,x[2][1]) returns the corresponding numerical values of the coefficients.

GraphMatFun.get_polynomial — Method

get_polynomial(graph::Compgraph)

Return the polynomial underlying the computational graph.

GraphMatFun.get_polynomial_coefficients — Method

get_polynomial_coefficients(graph::Compgraph)

Return the coefficients of the polynomial underlying the computational graph. The coefficients are expressed in the monomial basis and sorted from the leading coefficient to the constant term.

GraphMatFun.get_sorted_keys — Method

v=get_sorted_keys(graph)

Returns a list of all nodes, sorted.

GraphMatFun.get_topo_order — Method

(order,can_be_deallocated,max_nodes)=get_topo_order(graph; priohelp=Dict{Symbol,Float64}(),free_mem_bonus=1000,will_not_deallocate=[:I],input=:A)

Computes a topological sort of graph, that is, an ordering of the nodes following which the function the graph represents can be evaluated. The priohelp kwarg can be used to obtain a different topological ordering by changing the node priority. The code assumes that the nodes :I and input do not need to be computed.

The code uses a heuristic to minimize pathwidth. It is based on a point system. You can influence the computation order by providing a priohelp. If you want node :B4 to be computed earlier, you can set priohelp[:B4]=-5000.0. The free_mem_bonus is used in the heuristic to prioritize the computation of nodes which release other nodes. The vector will_not_deallocate influences the order specifying nodes that will not be deallocated and therefore gets no free_mem_bonus.

The return value order is a Vector of Symbols, and can_be_deallocated is a Vector{Vector{Symbol}} where the element i specifies the Symbols that are unused after step i in the ordering. The max_nodes is the pathwidth.

GraphMatFun.get_topo_order_degopt — Method

order=get_topo_order_degopt(k)

A special implementation of get_topo_order for degree optimal polynomials generated with graph_degopt. The natural order of computation is to compute row by row. See also get_degopt_crefs.

GraphMatFun.graph_bbc_exp — Method

(graph,cref)=graph_bbc_exp(k;T=Float64)

Computes a polynomial evaluation approximating the exponential using k matrix multiplications following the procedure in the reference. The coefficients are directly copied from the paper.

The evaluation is embedded in the degopt-format, and for k<3, the evaluation is using the Paterson–Stockmeyer method.

Reference:

Computing the matrix exponential with an optimized Taylor polynomial approximation, P. Bader, S. Blanes, and F. Casas, Mathematics, 7(12), 2019.

GraphMatFun.graph_bbcs_cheb_exp — Method

(graph,cref)=graph_bbcs_cheb_exp(k;T=Complex{BigFloat})

Computes a polynomial evaluation approximating the exponential for skew-Hermitian matrices using k matrix multiplications following the procedure in the reference.

For k > 5 it resorts to scaling-and-squaring of the k = 5 graph. The graph is in the degopt format.

Reference:

An efficient algorithm to compute the exponential of skew-Hermitian matrices for the time integration of the Schrödinger equation, P. Bader, S. Blanes, F. Casas, M. Seydaoglu, arXiv:2103.10132

GraphMatFun.graph_degopt — Method

(graph,crefs)=graph_degopt(k;T=ComplexF64,input=:A)
(graph,crefs)=graph_degopt(x,z;input=:A)
(graph,crefs)=graph_degopt(d::Degopt;input=:A)

Corresponds to the (for a fixed numer of multiplications) degree-optimal polynomial

B1=A
B2=(x *I+x *A)(x *I+x *A)
B3=(x *I+x *A+x*B2)(x *I+x *A+x *B2)
 ..

and

Out=z*I+z*A1+z*B2+z*B3+z*B4...

The x-values are given in the argument x, which is a Vector{Tuple{Vector,Vector}}, containing the elements of each sum. The z-vector contains the elements to form the output, and input determines the name of the matrix A above. If the parameter k is supplied instead of the coefficients, all coeffs will be set to one.

Reference: The general recursion is mentioned in equation (9) in this paper:

Computing the matrix exponential with an optimized Taylor polynomial approximation, P. Bader, S. Blanes, and F. Casas, Mathematics, 7(12), 2019.

GraphMatFun.graph_denman_beavers — Method

    (graph,cref)=graph_denman_beavers(k;T=ComplexF64,cref_mode=0)

Creates the graph corresponding to the Denman-Beavers iteration for the matrix square root, using k iterations. The kwarg cref_mode specifies if references to all X and Y (cref_mode=0) should be stored or just Y (cref_mode=-1) or just X (cref_mode=1).

Reference:

The matrix sign function and computations in systems, E. Denman, A. Beavers, Applied Mathematics and Computation, 2(1), 1976, https://doi.org/10.1016/0096-3003(76)90020-5

GraphMatFun.graph_exp_native_jl — Method

 (graph,crefs)=graph_exp_native_jl(A; input=:A)

Creates a graph for the native scaling-and-squaring for the matrix exponential, as implemented in Julia. The matrix A is taken as input to determine the length of the Padé approximant and the number of squares applied, as well as determining the type of the coefficients. The kwarg input determines the name of the matrix, in the graph.

References:

N. J. Higham. Functions of Matrices. SIAM publications, Philadelphia, PA, 2008.

N. J. Higham. The Scaling and Squaring Method for the Matrix Exponential Revisited. SIAM J. Matrix Anal. Appl., 2005 26:4, 1179-1193

Julia's matrix exponential, at the time of conversion: https://github.com/JuliaLang/julia/blob/697e782ab86bfcdd7fd15550241fe162c51d9f98/stdlib/LinearAlgebra/src/dense.jl#L554

GraphMatFun.graph_exp_native_jl_degopt — Method

 (graph,crefs)=graph_exp_native_jl_degopt(A; input=:A)

Same as graph_exp_native_jl but with calls to graph_degopt for contruction of numerator and denominator polynomials.

GraphMatFun.graph_horner — Method

 (graph,crefs)=graph_horner(a; input=:A, B=:B, C=:C, scaling=1.0)

Generates the graph for the polynomial using Horner's method. More precisely, it corresponds to the evaluation of the polynomial

p(s) = a[1] + a[2]*(αs) + ... + a[n-1]*(αs)^(n-2) + a[n]*(αs)^(n-1),

as the recursion

Cj=s*B{j+1}
Bj=a[j]*I+α*Cj,

where α=scaling.

The kwargs B and C specifies the base-names of these intermediate variables.

GraphMatFun.graph_horner_degopt — Method

 (graph,crefs)=graph_horner_degopt(a; scaling=1.0, input=:A)

Generates a polynomial using Horner's evaluation scheme. The polynomial

 p(s) = a[1] + a[2]*(αs) + ... + a[n-1]*(αs)^(n-2) + a[n]*(αs)^(n-1),

is evaluated as

 p(s) = a[1] + (αs)*(a[2] + (αs)*(... + (αs)*(a[n-1] + a[n]*(αs))...)),

where α=scaling. However, the function uses a call to graph_degopt, resulting in more degrees of freedom in crefs. See also graph_horner.

GraphMatFun.graph_monomial — Method

 (graph,crefs)=graph_monomial(a; input=:A, polyname=:P)

Generates the graph for the polynomial using the monomial basis coefficients. More precisely, it corresponds to the evaluation of the polynomial

p(s) = a[1] + a[2]*s + ... + a[n]*s^(n-1),

where s^k is naively evaluated as s^k=s*s^(k-1) for k=2,3,...,n-1. The kwarg polyname specifies the name of intermediate variables.

GraphMatFun.graph_monomial_degopt — Method

 (graph,crefs)=graph_monomial_degopt(a; input=:A)

Generates the same polynomial as graph_monomial, in the monomial basis. However, it does so by wrapping a call to graph_degopt, resulting in more degrees of freedom in crefs.

GraphMatFun.graph_newton_schulz — Function

 (graph,crefs)=graph_newton_schulz(k, T=ComplexF64; input=:A, B=:B, C=:C, V=:V)

Does k iterations of the Newton–Schulz iteration for approximating the inverse of a matrix A (name given by input), i.e., the recursion

V_k+1 = V_k*(2*I - A*V_k),

with V_0=A. The recursion is implemented using the graph-operations

Z_i=A*V_i
Q_i=2*I-Z_i
V_{i+1}=V_i*Q_i,

and the kwargs Z, Q, and V determines the naming of the corresponding intermediate steps.

References:

Günther Schulz. Iterative Berechnung der reziproken Matrix. Z. Angew. Math. Mech., 13:57–59, 1933.

N. J. Higham. Functions of Matrices. SIAM publications, Philadelphia, PA, 2008.

GraphMatFun.graph_newton_schulz_degopt — Function

 (graph,crefs)=graph_newton_schulz_degopt(k, T=ComplexF64; input=:A)

Does k iterations of the Newton–Schulz iteration for approximating the inverse, using the recursion

Z_i=A*V_i
V_{i+1}=V_i*(2*I-Z_i).

The function makes a call to graph_degopt, resulting in more degrees of freedom in crefs. See also graph_newton_schulz.

GraphMatFun.graph_ps — Method

 (graph,crefs)=graph_ps(a; input=:A,
                      B_base=:B, C_base=:C, P_base=:P)

Generates the graph for the Paterson–Stockmeyer procedure with monomial basis coefficieents. More precily, it corresponds to evaluation of the polynomial

p(s) = a[1] + a[2]*s + ... + a[n]*s^(n-1).

Reference:

On the number of nonscalar multiplications necessary to evaluate polynomials, M. Paterson, L. Stockmeyer, SIAM J. Comput., 2(1), 1973.

The code follows the description in:

Optimality of the Paterson–Stockmeyer method for evaluating matrix polynomials and rational matrix functions, M. Fasi, Linear Algebra Appl., 574, 2019.

GraphMatFun.graph_ps_degopt — Method

 (graph,crefs)=graph_ps_degopt(a; input=:A)

Generates the same polynomial as graph_ps, i.e., the graph for the Paterson–Stockmeyer procedure with monomial basis coefficieents. However, it does so by wrapping a call to graph_degopt, resulting in more degrees of freedom in crefs.

GraphMatFun.graph_rational — Method

 (graph, cref) = graph_rational(den_coeffs, num_coeffs, poly_gen=graph_ps)
 (graph, cref) = graph_rational(den_graph, den_graph; den_cref=Vector{Tuple{Symbol,Int}}(), num_cref=Vector{Tuple{Symbol,Int}}())

Generates the graph for the rational approximation

r(A)=q(A)^{-1}p(A)

where p(A) and q(A) are polynomials defined by the coeficients den_coeffs and num_coeffs, and generated by the function poly_gen, which is called as (graph,cref)=polygen(coeffs), see, e.g., `graphmonomialandgraph_ps`.

The alternative call-signature involves the graphs for p and q directly, as den_graph and num_graph. The corresponding den_cref and num_cref can also be passed to be modified accodringly, otherwise the return value cref is empty for this call.

GraphMatFun.graph_sastre_exp — Function

(graph,cref)=graph_sastre_exp(k,method=:auto)

Computes a polynomial evaluation approximating the exponential using k matrix multiplications following a method given in the reference. The schemes are embedded into degop-format.

Methods are:

:ps_degopt, Paterson–Stockmeyer method embedded into degopt-format.

:y1s, given by equations (34)-(35)

:z1ps, given by equations (34)-(35) and (52)

:h2m, given by equations (34)-(35) and (69)

Not all combinations of k and method are implemented. Available ones are:

k<3, method=:ps_degopt

k=3, method=:y1s, as per Table 4 in the reference

k=4, method=:y1s

k=6, method=:h2m, as per Table 11 in the reference

k=8, method=:z1ps, as per Table 7 in the reference

The default option method=:auto will choose method according to the value of k, as prescribed above.

Reference:

Efficient evaluation of matrix polynomials, J. Sastre. Linear Algebra and its Applications ,Volume 539, 2018, Pages 229-250, https://doi.org/10.1016/j.laa.2017.11.010

GraphMatFun.graph_sastre_poly — Method

(graph,cref)=graph_sastre_poly(b)

Computes the degree-8 polynomial

p(z)=b[1]+z*b[2]+z^2*b[3]+...+z^8*b[9]

according to Example 3.1 in the reference.

Reference:

Efficient evaluation of matrix polynomials, J. Sastre. Linear Algebra and its Applications ,Volume 539, 2018, Pages 229-250, https://doi.org/10.1016/j.laa.2017.11.010

GraphMatFun.graph_sastre_yks_degopt — Method

(graph,cref)=graph_sastre_yks_degopt(k,s,c)

Transforms the polynomial evaluation format given by equations (62)-(65) in the reference to degop-format. The graph is a representation of y_{k s}. Input c is a grouping of the coefficients as given by the representation (62)-(65). c is a Vector{Vector{Vector}} of length k+1, representing

[
[c_i^{(0,1)}, c_i^{(0,2)}]
[c_i^{(1,1)}, c_i^{(1,2)}, c_i^{(1,3)}, c_i^{(1,4)}, c_i^{(1,5)}, c_i^{(1,6)}]
...
[c_i^{(k,1)}, c_i^{(k,2)}, c_i^{(k,3)}, c_i^{(k,4)}, c_i^{(k,5)}, c_i^{(k,6)}]
]

Hence, c[1] contains two vectors, the first of length s-1 and the second of length s. (Note: In the first vector the constant for I is set to zero) and c[2] up to c[k+1] conatins six vectors: Even-numbered vectors have length s+1 Odd-numbered vectors have length j-1, where j is the intex in c, e.g., c[2][1] has one element and c[3][5] have two. For example, equations (57)-(59) are implemented as:

c = [
[[c15, c16], [0.0, 0, 0]], # (57)
[[1.0], [0, c13, c14], [1.0], [c11, 0, c12], [c10], [0.0, 0, 0]], # (58)
[[0.0, 1], [0, c8, c9], [c7, 1], [0, c6, 0], [c4, c5], [c1, c2, c3]], # (59)
]

Reference:

Efficient evaluation of matrix polynomials, J. Sastre. Linear Algebra and its Applications ,Volume 539, 2018, Pages 229-250, https://doi.org/10.1016/j.laa.2017.11.010

GraphMatFun.graph_sid_exp — Method

(graph,cref)=graph_sid_exp(k;T=Float64)

Computes a polynomial evaluation approximating the exponential using k matrix multiplications following the procedure in the reference. The coefficients are directly copied from the paper.

The evaluation is embedded in the degopt-format, using the function graph_sastre_yks_degopt. Moreover, for k<=3 it uses the Paterson–Stockmeyer method.

Reference:

Boosting the computation of the matrix exponential, J. Sastre, J. Ibáñez, E. Defez, Appl. Math. Computation, 340, 2019, 206-220.

GraphMatFun.grow! — Method

grow!(degopt::Degopt)

Increases the degopt by one multiplication without modifying the function values.

GraphMatFun.has_identity_lincomb — Method

has_identity_lincomb(graph) -> Bool

Checks whether the graph has a node which is a linear combination of identity matrices.

GraphMatFun.has_trivial_nodes — Method

has_trivial_node(graph) -> Bool

Checks whether the graph has trivial nodes, that is, multiplications by the identity or linear systems whose coefficient is the identity matrix.

GraphMatFun.import_compgraph — Method

graph=import_compgraph(fname)

Reads a graph stored in the computation graph format (cgr) in the file fname.

GraphMatFun.merge_graphs — Method

graph=merge_graphs(graph1,graph2;prefix1="",prefix2="G2",skip_basic1=true,skip_basic2=true,cref1=Vector(),cref2=Vector(),input1=:A,input2=:A)

Takes all the nodes and edges in graph1 and graph2 and generates a new graph. The node names are in graph1 are changed by adding a prefix prefix1 and graph2 correspondingly. The nodes :I and :A are unchanged if skip_basicX=true. All coefficient references crefX are modified accordingly.

GraphMatFun.opt_gauss_newton! — Method

(iter,resnorm)=opt_gauss_newton!(graph, objfun, discr; maxit = 100, logger = 0,
                                 errtype = :abserr, stoptol = 1e-6,
                                 input=:A, cref = get_all_cref(graph),
                                 γ0 = 1.0, linlsqr = :backslash, droptol = 0)

Applies the Gauss–Newton algorithm to solve the nonlinear least squares problem: Fit the output of the graph to the values of objfun, in the points discr.

The variable graph is modified during the iterations. The function returns the iteration iter where it terminated, and the corresponding residual norm resnorm.

The kwargs are as follows: maxit determines the maximum number of iterations. If logger has a value >0, then the intermediate results are printed. errtype determines the error type measured, see adjust_for_errtype!. stoptol is the corresponding stopping tolerence. cref is a Vector{Tuple{Symbol,Int}} that determines which coefficients of graph that are considered free variables and optimized. input is the label corresponding to the input node of the graph. The stepsize can be scaled with γ0. linlsqr and droptol determines how the inner linear least squares problem is solved; see solve_linlsqr!.

GraphMatFun.opt_linear_fit! — Method

opt_linear_fit!(graph, objfun, discr, linear_cref;
                input = :A,
                errtype = :abserr,
                linlsqr = :backslash,
                droptol = 0)

Linear fitting of a graph of the form

c_1 g_1(x) + c_2 g_2(x) + … + c_n g_n(x)

to the values of objfun, in the points discr. Reference to the coefficients c_1,…,c_n should be given linear_cref.

The variable graph is modified during the iterations and the function has no return value.

See opt_gauss_newton! for a description the kwarg errtype and input, and solve_linlsqr! for the kwargs linlsqr anddroptol`.

GraphMatFun.rename_node! — Function

rename_node!(graph,src,dest,cref=Vector())

This changes the name of the node src to dest and updating all references to the node including coefficient references in cref.

GraphMatFun.scale! — Method

scale!(degopt::Degopt,α)

Effectively change a Degopt such that the input is scaled by α. If p is the original function, p(α x) will be the new function.

GraphMatFun.set_coeffs! — Function

set_coeffs!(graph, x, cref=get_all_cref(graph))

Sets the coefficient values in the coefficients specified in cref::Vector to the values in the vector in x::Vector.

GraphMatFun.solve_linlsqr! — Method

d = solve_linlsqr!(A, b, linlsqr, droptol)

Solves the linear least squares problem

Ad=b.

The argument linlsqr determines how the linear least squares problem is solved. It can be :backslash, :real_backslash, :nrmeq, :real_nrmeq, :svd, or :real_svd. For the latter two options singular values below droptol are disregarded. The :real_X options optimizes d in the space of real vectors. The input matrix A is sometimes overwritten.

GraphMatFun.square! — Method

square!(degopt::Degopt)

Effectively square a Degopt in the sense that the output is square. If p is the original function, p(x)^2 will be the new function.

LinearAlgebra.normalize! — Function

normalize!(degopt,tp=:row1)

Normalizes the degopt coefficients, in the way specified by tp. If the rp==:row1 the degopt will be transformed to an equivalent degopt with first row equal to (0 1) (0 1).