Code generation

GraphMatFun.gen_code — Function

gen_code(fname,graph; priohelp=Dict{Symbol,Float64}(),
         lang=LangJulia(),funname="dummy",precomputed_nodes=[:A])

Generates the code for the graph in the language specified in lang and writes it into the file fname. The string funname is the function name. Topological order of the nodes is comptued using get_topo_order and priohelp can be used to influence the order. The nodes listed in precomputed_nodes are viewed as inputs, and code to compute these nodes are not computed.

Currently supported languages: LangC_MKL, LangC_OpenBLAS, LangJulia, LangMatlab.

GraphMatFun.LangJulia — Type

LangJulia(overwrite_input=true,inline=true,dot_fusing=true)

Code generation in julia language, with optional overwriting of input, inlining the function and optional usage of dot fusion.

GraphMatFun.LangMatlab — Type

LangMatlab()

Code generation for the Matlab language.

GraphMatFun.LangC_MKL — Type

 LangC_MKL(gen_main::Bool)

Code generation in C using MKL implementation of BLAS.

GraphMatFun.LangC_OpenBLAS — Type

 LangC_OpenBLAS(gen_main::Bool)

Code generation in C using OpenBLAS implementation of BLAS.