Graphs and their manipulation

Manipulating graph topology

GraphMatFun.Compgraph — Type

graph=Compgraph(T::Type=ComplexF64)

Creates an empty computation graph of with coefficients of type T.

graph=Compgraph(T,orggraph::Compgraph)

Converts a graph such that the coefficients become type T. Note that T can be a Number but can be other objects with defined operations, or just Any.

GraphMatFun.add_mult! — Function

add_mult!(graph,node,p1,p2)

Adds a multiplication of node p1 and p2 to the graph. The result is stored in node node.

GraphMatFun.add_lincomb! — Function

add_lincomb!(graph,node,α1,p1,α2,p2)

The operation α1*p1+α2*p2 is added to the graph. The result is stored in node node.

GraphMatFun.add_ldiv! — Function

add_ldiv!(graph,node,p1,p2)

The operation inv(p1)*p2 is added to the graph. The result is stored in node node.

GraphMatFun.add_sum! — Function

cref=add_sum!(graph,node,c,nodelist,base_name=node)

Adds a linear combination of more than two nodes, given in nodelist::Vector, with coefficients given in c. The base_name is temporary variables for the summing. The sum is stored in node.

Returns cref list with references.

GraphMatFun.add_output! — Function

add_output!(graph,node)

Adds an output node to the graph.

GraphMatFun.del_output! — Function

del_output!(graph,node)

Removes the output node from the list of output nodes of the graph. This function does not remove the node from the graph.

GraphMatFun.clear_outputs! — Function

clear_outputs!(graph)

Clears the list of output nodes of the graph. This function does not remove the node nodes from the graph.

GraphMatFun.rename_node! — Function

rename_node!(graph,src,dest,cref=Vector())

This changes the name of the node src to dest and updating all references to the node including coefficient references in cref.

GraphMatFun.del_node! — Function

del_node!(graph,node)

Deletes a node from the graph, and all data associated with it, except for graph.outputs.

Evaluating of graph

GraphMatFun.eval_graph — Function

result=eval_graph(graph, x; vals=nothing,
                  input=:A, output=size(graph.outputs,1),
                  comporder=nothing)

Evaluates a graph in the value x which is typically a scalar value or a matrix. If x is a Vector, the values will be evaluated elementwise.

The comporder is a Vector of nodes specifying in which order the graph should be computed. By default get_topo_order is used.

The output is an Int specifying which node should be considered as output. The output node is graph.outputs[output].

The vals is used to inspect contents other than the output inside the graph. Typically vals is a Dict. It will be modified to contain the computed nodes of the graph. If we wish to inspect node X3, we can initiate an empty dict as input:

julia> vals=Dict{Symbol,Any}();
julia> eval_graph(graph,A,vals=vals);
julia> vals[:X3]

GraphMatFun.get_topo_order — Function

(order,can_be_deallocated,max_nodes)=get_topo_order(graph; priohelp=Dict{Symbol,Float64}(),free_mem_bonus=1000,will_not_deallocate=[:I],input=:A)

Computes a topological sort of graph, that is, an ordering of the nodes following which the function the graph represents can be evaluated. The priohelp kwarg can be used to obtain a different topological ordering by changing the node priority. The code assumes that the nodes :I and input do not need to be computed.

The code uses a heuristic to minimize pathwidth. It is based on a point system. You can influence the computation order by providing a priohelp. If you want node :B4 to be computed earlier, you can set priohelp[:B4]=-5000.0. The free_mem_bonus is used in the heuristic to prioritize the computation of nodes which release other nodes. The vector will_not_deallocate influences the order specifying nodes that will not be deallocated and therefore gets no free_mem_bonus.

The return value order is a Vector of Symbols, and can_be_deallocated is a Vector{Vector{Symbol}} where the element i specifies the Symbols that are unused after step i in the ordering. The max_nodes is the pathwidth.

Linear combinations and coefficients

GraphMatFun.get_all_cref — Function

v=get_all_cref(graph)

Returns a list with references to all coefficients in the graph. The list is sorted.

GraphMatFun.get_coeffs — Function

x=get_coeffs(graph, cref=get_all_cref(graph))

Gets the coefficient values for the coefficients specified in cref::Vector.

GraphMatFun.set_coeffs! — Function

set_coeffs!(graph, x, cref=get_all_cref(graph))

Sets the coefficient values in the coefficients specified in cref::Vector to the values in the vector in x::Vector.

GraphMatFun.extract_sums — Function

v=extract_sums(graph)

::Vector{Tuple{Vector{Float64},Vector{Symbol},Vector{Symbol}}}`

Returns a representation of sums in graph which may potentially be merged. The vector sums contains a tuple for each of these sums. The three entries of the tuple are:

a vector of Float64 values that represent the coefficients of the summands;

a vector of Symbols that correspond to the summands; and

a vector of intermediate Symbols (i.e, nodes) that can be merged.

The first two vectors have the same number of entries, one for each element that can be merged in the sum.

GraphMatFun.get_sorted_keys — Function

v=get_sorted_keys(graph)

Returns a list of all nodes, sorted.

Reading and writing to file

GraphMatFun.export_compgraph — Function

export_compgraph(graph, fname; main_output=nothing,
                 order=get_topo_order(graph)[1],
                 fun="", dom="", err="", genby="",
                 descr="",
                 user=current user)

Exports the graph to the file fname using the computation graph format (cgr). These kwargs are strings or values that are stored as comments in the file

descr of the graph

fun function it approximates

dom domain it approximates

err error in this domain

genby script that generated this file

user a username that created the file

These kwarg influence how the graph is stored:

order specifies an order the graph nodes are stored

main_output is a Symbol specifying what is the output.

CGR format:

Every line corresponds to a node / operation. The syntax is matlab compatible, although gen_code produces faster matlab code.

See also export_compgraph.

GraphMatFun.import_compgraph — Function

graph=import_compgraph(fname)

Reads a graph stored in the computation graph format (cgr) in the file fname.

See also import_compgraph.